move subscription tests to ts
