+++ tests/EvmAcceptanceTests/test/subscriptions/WithTwoSubscribers.ts
@@ -1,21 +1,27 @@
-const {expect} = require("chai");
-const {ethers} = require("hardhat");
+import deepEqualInAnyOrder from "deep-equal-in-any-order";
+import chai from "chai";
 
-const general_helper = require("../../helper/GeneralHelper");
-const parallelizer = require("../../helper/Parallelizer");
+chai.use(deepEqualInAnyOrder);
+
+import { expect } from "chai";
+import { Contract} from "ethers";
+import hre, { ethers } from "hardhat";
+import parallelizer from "../../helper/Parallelizer";
+import { Event, waitForEvents } from "./shared";
+import { WebSocketProvider } from "@ethersproject/providers";
 
 describe("Subscriptions functionality", function () {
-  let contract;
-  let eventsContract;
-  let provider;
-  let senderAddress;
+  let contract: Contract;
+  let eventsContract: Contract;
+  let senderAddress: string;
+  let provider: WebSocketProvider
   before(async function () {
     contract = await parallelizer.deployContract("Subscriptions");
-    senderAddress = contract.signer.address;
+    senderAddress = await contract.signer.getAddress();
   });
 
   beforeEach(async function () {
-    provider = new ethers.providers.WebSocketProvider(general_helper.getWebsocketUrl());
+    provider = new ethers.providers.WebSocketProvider(hre.getWebsocketUrl());
     eventsContract = new ethers.Contract(contract.address, contract.interface, provider);
   });
 
@@ -23,44 +29,38 @@ describe("Subscriptions functionality", function () {
     eventsContract.removeAllListeners();
   });
 
-  // Ensures all events in current eventloop run are dispatched
-  async function waitForEvents(events, timeout = 5000) {
-    await new Promise((r) => setTimeout(r, timeout));
-    return events;
-  }
-
   describe("When two subscribers listen to events", function () {
     it("Should receive an event coming only from contract it is subscribed to", async function () {
       const secondContract = await parallelizer.deployContract("Subscriptions");
-      const secondProvider = new ethers.providers.WebSocketProvider(general_helper.getWebsocketUrl());
+      const secondProvider = new ethers.providers.WebSocketProvider(hre.getWebsocketUrl());
       const secondEventsContract = new ethers.Contract(
         secondContract.address,
         secondContract.interface,
         secondProvider
       );
 
-      let receivedEvents = [];
-      let secondContractReceivedEvents = [];
+      let receivedEvents: Event[] = [];
+      let secondContractReceivedEvents: Event[] = [];
 
       const filter = eventsContract.filters.Event0();
       eventsContract.on(filter, (from, to, amount, _event) => {
-        receivedEvents.push([from, to, amount]);
+        receivedEvents.push({from, to, amount});
       });
 
       const filter2 = secondEventsContract.filters.Event0();
       secondEventsContract.on(filter2, (from, to, amount, _event) => {
-        secondContractReceivedEvents.push([from, to, amount]);
+        secondContractReceivedEvents.push({from, to, amount});
       });
 
       const VECTORS = [
-        [senderAddress, "0x05A321d0B9541Ca08d7e32315Ca186cC67A1602c", new ethers.BigNumber.from(100)],
-        [senderAddress, "0xF0Cb24aC66ba7375Bf9B9C4Fa91E208D9EAAbd2e", new ethers.BigNumber.from(200)]
+        {from: senderAddress, to: "0x05A321d0B9541Ca08d7e32315Ca186cC67A1602c", amount: ethers.BigNumber.from(100)},
+        {from: senderAddress, to: "0xF0Cb24aC66ba7375Bf9B9C4Fa91E208D9EAAbd2e", amount: ethers.BigNumber.from(200)}
       ];
 
       for (let i = 0; i < VECTORS.length; ++i) {
-        const [_SENDER, DEST, AMOUNT] = VECTORS[i];
-        await expect(contract.event0(DEST, AMOUNT)).to.emit(contract, "Event0");
-        await expect(secondContract.event0(DEST, AMOUNT)).to.emit(secondContract, "Event0");
+        const event: Event = VECTORS[i];
+        await expect(contract.event0(event.to, event.amount)).to.emit(contract, "Event0");
+        await expect(secondContract.event0(event.to, event.amount)).to.emit(secondContract, "Event0");
       }
       receivedEvents = await waitForEvents(receivedEvents);
       secondContractReceivedEvents = await waitForEvents(secondContractReceivedEvents);
@@ -70,27 +70,27 @@ describe("Subscriptions functionality", function () {
     it("Should deliver event to both", async function () {
       const secondEventsContract = new ethers.Contract(contract.address, contract.interface, provider);
 
-      let receivedEvents = [];
-      let secondContractReceivedEvents = [];
+      let receivedEvents: Event[] = [];
+      let secondContractReceivedEvents: Event[] = [];
 
       const filter = eventsContract.filters.Event0();
       eventsContract.on(filter, (from, to, amount, _event) => {
-        receivedEvents.push([from, to, amount]);
+        receivedEvents.push({from, to, amount});
       });
 
       const filter2 = secondEventsContract.filters.Event0();
       secondEventsContract.on(filter2, (from, to, amount, _event) => {
-        secondContractReceivedEvents.push([from, to, amount]);
+        secondContractReceivedEvents.push({from, to, amount});
       });
 
       const VECTORS = [
-        [senderAddress, "0x05A321d0B9541Ca08d7e32315Ca186cC67A1602c", new ethers.BigNumber.from(100)],
-        [senderAddress, "0xF0Cb24aC66ba7375Bf9B9C4Fa91E208D9EAAbd2e", new ethers.BigNumber.from(200)]
+        {from: senderAddress, to: "0x05A321d0B9541Ca08d7e32315Ca186cC67A1602c", amount: ethers.BigNumber.from(100)},
+        {from: senderAddress, to: "0xF0Cb24aC66ba7375Bf9B9C4Fa91E208D9EAAbd2e", amount: ethers.BigNumber.from(200)}
       ];
 
       for (let i = 0; i < VECTORS.length; ++i) {
-        const [_SENDER, DEST, AMOUNT] = VECTORS[i];
-        await expect(contract.event0(DEST, AMOUNT)).to.emit(contract, "Event0");
+        const event: Event = VECTORS[i];
+        await expect(contract.event0(event.to, event.amount)).to.emit(contract, "Event0");
       }
 
       receivedEvents = await waitForEvents(receivedEvents);
@@ -103,30 +103,30 @@ describe("Subscriptions functionality", function () {
     xit("Should deliver event to only a valid one", async function () {
       const secondEventsContract = new ethers.Contract(contract.address, contract.interface, provider);
 
-      let receivedEvents = [];
-      let secondContractReceivedEvents = [];
+      let receivedEvents: Event[] = [];
+      let secondContractReceivedEvents: Event[] = [];
 
       const filter = eventsContract.filters.Event0();
       eventsContract.on(filter, (from, to, amount, _event) => {
-        receivedEvents.push([from, to, amount]);
+        receivedEvents.push({from, to, amount});
       });
 
       const filter2 = secondEventsContract.filters.Event0();
       secondEventsContract.on(filter2, (from, to, amount, _event) => {
-        secondContractReceivedEvents.push([from, to, amount]);
+        secondContractReceivedEvents.push({from, to, amount});
       });
 
       // Deliberately close subscription
       secondEventsContract.removeAllListeners();
 
       const VECTORS = [
-        [senderAddress, "0x05A321d0B9541Ca08d7e32315Ca186cC67A1602c", new ethers.BigNumber.from(100)],
-        [senderAddress, "0xF0Cb24aC66ba7375Bf9B9C4Fa91E208D9EAAbd2e", new ethers.BigNumber.from(200)]
+        {from: senderAddress, to: "0x05A321d0B9541Ca08d7e32315Ca186cC67A1602c", amount: ethers.BigNumber.from(100)},
+        {from: senderAddress, to: "0xF0Cb24aC66ba7375Bf9B9C4Fa91E208D9EAAbd2e", amount: ethers.BigNumber.from(200)}
       ];
 
       for (let i = 0; i < VECTORS.length; ++i) {
-        const [_SENDER, DEST, AMOUNT] = VECTORS[i];
-        await expect(contract.event0(DEST, AMOUNT)).to.emit(contract, "Event0");
+        const event: Event = VECTORS[i];
+        await expect(contract.event0(event.to, event.amount)).to.emit(contract, "Event0");
       }
       receivedEvents = await waitForEvents(receivedEvents);
       secondContractReceivedEvents = await waitForEvents(secondContractReceivedEvents);
